pyproject.toml
[tool.poetry]
name = "shop-tms-drf"
version = "0.1.0"
description = ""
authors = ["Alexandr <maximchik.alexandr@yandex.ru>"]
readme = "README.md"
packages = [{include = "shop_tms_drf"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
django-environ = "^0.10.0"
prospector = "^1.9.0"
djoser = "^2.1.0"
django-seed = "^0.3.1"
faker = "^18.3.1"
faker-commerce = "^1.0.3"
pytest-django = "^4.5.2"
celery = "^5.2.7"
redis = "^4.5.4"
django-celery-beat = "^2.5.0"
django-rest-swagger = "2.2.0"
drf-yasg = "1.21.4"
drf-yasg2 = "1.19.1"
aiogram = "^2.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
